Initialize typescript-nextjs-sass app
There are just few things specific. In _app.tsx, which is a kind of main frame of next.js app, I placed <div id={modalContainerID}>. It is for initializing basic structure including modal. So I added @types/react-dom dependency bc I need to get reference of ReactDOM to call createPortal when I make a modal.

And the directory structure. I write some on README.md, but it can be changed depending on actual pages that the designer's gonna give. Probabaly if the feature received is pretty complex, then we should split or make more directories like widgets, templates ... we should talk.